Use ioctl for time update if /dev/alarm is present.
Sony stock kernels <3.10 require to set the system time on boot via
the alarm framework. The ioctl will seemingly fail as the kernel
cannot write to the rtc, but it still sets the system time correctly.
Setting the system in this way is needed to avoid that timers in
the alarm framework would still count from some date in the 1970's.
One instance of problems caused by that is failing auto-reconnect
for wifi network, where a scan triggered by the android framework
never finishes, an thus an internal lists of wifi networks is
never filled.

Change-Id: If03e12128ad66998adab204afce1db9bb8938770
Signed-off-by: Alexander Diewald <Diewi@diewald-net.com>